Fixed user data does not exist in profile, then change background of forgot password to white
